feishu-print-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
slowapi==0.1.9
openai>=1.0.0
httpx>=0.24.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
